Move helpers from CiviUnitTestCase to DbTestTrait
* `assertDBState(...)`
* `assertDBNotNull(...)`
* `assertDBNull(...)`
* `assertDBRowNotExist(...)`
* `assertDBRowExist(...)`
* `assertDBCompareValue(...)`
* `assertDBCompareValues(...)`
* `assertDBQuery(...)`
